refactor : Dynamic Routing
parameter -> path로 변경
