+++ src/components/profile/bid/SuccessBidItem.tsx
@@ -24,10 +24,7 @@ export default function SuccessBidItem({
     <SuccessBidItemTag
       {...rest}
       onClick={() => {
-        router.push({
-          pathname: '/auction/view',
-          query: { id: biddingItem.id },
-        });
+        router.push('/auction/' + biddingItem.id);
       }}
     >
       <article className="relative h-[6.25rem] w-[5.125rem] overflow-hidden rounded">
