+++ src/pages/auction/bidding/[id].tsx
